backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
PyJWT
python-dotenv
passlib
